publish pagination vendor files
Using php artisan vendor:publish allows to take control of the files
initially in the vendor folder by copying them in our part of the app.
It is useful when we want to tweak some code that we haven't written
ourselves.

Also added withQueryString method to search on categories without
breaking pagination.

Section 8 lesson 44 of laracasts from scratch tuto